Rename "thirdparty" to "lib".
