+++ components/worker_manager/worker.js
@@ -1,4 +1,4 @@
-importScripts('../../thirdparty/diff_match_patch.js');
+importScripts('../../lib/diff_match_patch.js');
 
 tasks = {
     diff: function(config){
