+++ lib/diff_match_patch.js
